requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
asyncio==3.4.3
colorama==0.4.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
alembic==1.13.3
# OCR Dependencies - Phase 3
pytesseract==0.3.10
Pillow>=10.0.0
opencv-python-headless>=4.8.0